api/requirements.txt
fastapi==0.110.1
uvicorn==0.28.0
fastmcp==0.5.1
pydantic==2.7.1
psycopg2-binary==2.9.9
sentence-transformers==2.5.1
GitPython==3.1.43
python-multipart==0.0.9
llama-index==0.10.12
llama-index-vector-stores-postgres==0.1.3
llama-index-readers-file==0.1.4
markdown==3.5.2
nltk==3.8.1